+++ CHANGES.rst
@@ -31,6 +31,9 @@ Changelog
 - Added transition icons
   [sgeulette]
 
+- Added batch action to change recipient groups
+  [sgeulette]
+
 - Improvements: disable own delete on contacts, block parent portlets on contacts, add local roles for dir_general
                 on contacts, corrected disabled treating_groups bug, updated voc cache, corrected transition batch
                 action, added task parent on task view, added method to test user group membership, ordered css